Added some features to the animator and polished the UI
yeah it's pretty cool, pretty rad if I do say to myself and it's actually really nice that you're reading this little commit-message and shows that you're a person with good karma! Hope you live long and don't become a dung-beetle or whatever due to a bad caste or whatever that Hindu cock-slammering bullwash is and if you're wondering I am totally not just writing whatever I come up with and the fact that you're insinuating that I put no effort into this message makes me actually hope now that you become a dung-beetle you hogwash piece of garbage.