Ok so you can beat the shit out of Hersh in TURN BASED MODE now!
Put that shit on the box my man, you can absolutely wail on a lobotomized man that can do nothing to defend himself. Believe men when I say that it's for the best as no-one wants to live in a time loop where a tired programmer beats you to death over and over again... No siree!