Entities now rotate to look when they're attacking and smaller changes.
Basically, fixed up so that when you and enemies attack you rotate to look at them. Along with this I am diving into the dialogue system and I've been looking at how the data is created and altered some things to make it easier to store and find narrative threads. Then it's just to load them into an NPC and to correctly display the data.